Added Jest
Configure webpack to make Jest snapshot testing works.

Remaining issue is how to do snapshot testing with the help of Avoiraz. Currently as a workaround/remporary but might be permanent solution is to use vue-server-renderer.

However the snapshot still contains Vuetify tags, instead of native HTML tags, and has the attribute data-server-rendered="true"